fix: refactor form components
