+++ src/components/form/button/button.css
@@ -1,0 +1,33 @@
+.button {
+  width: 100%;
+  height: 50px;
+  font-size: 18px;
+  line-height: 22px;
+  color: white;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: #000;
+  border-radius: 2px;
+  border: none;
+  transition: visibility 0s, background 0.3s;
+  cursor: pointer;
+}
+
+.button:hover {
+  background: rgba(0, 0, 0, 0.8);
+}
+
+.button:disabled {
+  color: #B8B8B8;
+  background: #EDEDED;
+  pointer-events: none;
+  user-select: none;
+}
+
+.button_inactive {
+  color: #B8B8B8;
+  background: #EDEDED;
+  pointer-events: none;
+  user-select: none;
+}
