+++ src/components/form/form-section/form-section.css
@@ -1,0 +1,4 @@
+.form__section {
+  margin: 0 0 32px 0;
+  position: relative;
+}
