+++ src/components/form/form.css
@@ -1,96 +1,3 @@
 .form {
   font-family: 'Inter', -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
 }
-
-.form__section {
-  margin: 0 0 32px 0;
-  position: relative;
-}
-
-.form__label {
-  display: block;
-  font-size: 22px;
-  font-weight: 500;
-  margin: 0 0 3px 0;
-}
-
-.form__label-text {
-  display: block;
-  margin-bottom: 4px;
-}
-
-@media (min-width: 450px) {
-  .form__label {
-    font-size: 16px;
-  }
-}
-
-.form__input {
-  border: 1px solid #ccc;
-  font-size: 16px;
-  padding: 16px;
-  width: 100%;
-  box-sizing: border-box;
-}
-
-.form__input-error {
-  font-size: 12px;
-  line-height: 15px;
-  color: #e53e3e;
-  opacity: 0;
-  position: absolute;
-  top: calc(100% + 5px);
-  left: 0;
-  transition: opacity 0.3s;
-}
-
-.form__input-error_active {
-  opacity: 1;
-}
-
-.form__toggle-password {
-  font-size: 16px;
-  background: none;
-  border: none;
-  cursor: pointer;
-  padding: 0;
-  margin: 0;
-  position: absolute;
-  top: 0;
-  right: -2px;
-}
-
-.form__submit {
-  width: 100%;
-  height: 50px;
-  font-size: 18px;
-  line-height: 22px;
-  color: white;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: #000;
-  border-radius: 2px;
-  border: none;
-  transition: visibility 0s, background 0.3s;
-  margin-top: 48px;
-  cursor: pointer;
-}
-
-.form__submit:hover {
-  background: rgba(0, 0, 0, 0.8);
-}
-
-.form__submit:disabled {
-  color: #B8B8B8;
-  background: #EDEDED;
-  pointer-events: none;
-  user-select: none;
-}
-
-.form__submit_inactive {
-  color: #B8B8B8;
-  background: #EDEDED;
-  pointer-events: none;
-  user-select: none;
-}
