+++ src/components/form/label/label.css
@@ -1,0 +1,17 @@
+.label {
+  display: block;
+  font-size: 22px;
+  font-weight: 500;
+  margin: 0 0 3px 0;
+}
+
+.label-text {
+  display: block;
+  margin-bottom: 4px;
+}
+
+@media (min-width: 450px) {
+  .label {
+    font-size: 16px;
+  }
+}
