+++ src/components/form/password-input/toggle-password-button/toggle-password-button.css
@@ -1,0 +1,11 @@
+.form__toggle-password {
+  font-size: 16px;
+  background: none;
+  border: none;
+  cursor: pointer;
+  padding: 0;
+  margin: 0;
+  position: absolute;
+  top: 0;
+  right: -2px;
+}
